requirements.txt
python-telegram-bot>=22.0
youtube-transcript-api
pytubefix
elevenlabs
anthropic
edge-tts
python-dotenv
cryptography
mem0ai
langchain-neo4j
rank-bm25
fastapi
uvicorn
jinja2
pytest
pytest-mock
pytest-asyncio
